+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from motivationalboost import create_app
+from motivationalboost.flask_app import create_app
 from mbconfig import MBConfig
